refactor(exceptions): remove exceptions package to match conventions
