--- src/main/java/io/github/spinel/exceptions/README.md
@@ -1,7 +1,0 @@
-# package `io.github.spinel.exceptions`
-
-## About
-
-This package obviously includes all the exceptions used in the Spinel engine.
-
-_Missing or incomplete documentation : you can contribute to this file on the project's [repository](https://github.com/l3alr0g/Spinel)_
